Repo Type changed to Long and Created Exceptions Along with Security Configuration

+++ src/main/java/com/cg/repo/BusBookingRepo.java
@@ -7,6 +7,6 @@ import com.cg.entity.Customer;
 
 
 @Repository
-public interface BusBookingRepo extends JpaRepository<Customer, Integer> {
+public interface BusBookingRepo extends JpaRepository<Customer, Long> {
 
 }
